timer: add generic timer API
Add a generic timer API that timer drivers can hook into.

The API is simple.

The BSP or driver creates an instance of timer_ops_t, fills in the driver specific
data, then calls timer_init().

After that, mdelay() or udelay() can be called to delay execution for
the specified time (milliseconds or microseconds, respectively).

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>