sp804: add driver
Add a driver for the ARM SP804 dual timer.

This driver will only use the first timer, called timer 1 in the
documentation.

To use this driver, the BSP must provide three constants:

	SP804_BASE	base address of the timer
	SP804_CLKDIV	clock divider
	SP804_CLKMULT	clock multiplier

The BSP is responsible for calling sp804_init(). The SP804 driver
creates a timer_ops_t and calls the generic timer_init().

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>