Juno: Detect SCP version incompatibility
There has been a breaking change in the communication protocols used
between the AP cores and the SCP. This requires to update both the
AP Trusted Firmware and the SCP firmware at the same time.

In case the user forgot to update the SCP ROM firmware, this patch
detects when it still uses the previous version of the communication
protocol. It will then output a comprehensive error message that helps
trouble-shoot the issue.

Change-Id: I7baf8f05ec0b7d8df25e0ee53df61fe7be0207c2